updated user flow by changing route paths
