+++ src/components/CaptureImage/CaptureImage.jsx
@@ -127,9 +127,6 @@ function CaptureImage() {
         <button onClick={closePhoto}>CLOSE!</button>
       </div>
       <br />
-      <div className="next-container">
-        <button onClick={goToSee}>Go To See</button>
-      </div>
     </section>
   );
 }
